Replace standalone statsd with the collectd plugin
This commit removes the standalone statsd server written in nodejs and
replaces it with a collectd plugin written in C.

This means that statsd metrics will now be sent via the local collectd
instance on each machine to the central collectd server, instead of
sending them via the local statsd instance to the graphite server.

nodejs is still required on the application servers because of other
dependencies, however we now can remove it from a number of other
machines.

I've tried to keep the configuration of the statsd plugin as close as
possible to our previous statsd standalone configuration. The one change
to be aware of is that collectd flushes its metrics by default every 10
seconds rather than the 5-second default in standalone statsd, so I have
adjusted our Graphite storage schemas accordingly.

We'll lose some resolution on our statsd metrics (10s instead of 5s),
but I don't expect significant issues from this.