Feature-flag standalone statsd
Add a feature flag that determines whether we use the standalone statsd
server or the collectd statsd plugin (which provides the same
functionality).

I've added the feature flag as separately, in this commit, so that it
can easily be reverted once we decide to move permanently to the
collectd statsd plugin in all environments.