Move mongodb module to mongodb_legacy namespace
We wish to start using the Puppetlabs mongodb module which shares the
name of our current inhouse module. Naming is hard: when we switch to
the external module then we should reserve the govuk_mongodb for any of
our own stuff that is used on top of the module (which is consistent in
the way we do things), so I used the "mongodb_legacy" namespace to make
it clear that we wish to move away from it.