Run unattended-upgrades on the dev VM as part of the puppet run
When we run govuk_puppet on a dev VM we expect that it will update all
the managed software on the machine, however for package updates it
relies on unattended-upgrades to actually perform the updates.  These
are set to run at some small hour in the morning, but most dev VMs are
not powered on at this time, so the updates don't actually run.

We noticed this when postgres 9.3.14 came out which introduced a breaking
change to the `\connect` line in sql dumps.  The servers were upgraded
but most dev VMs weren't and so replicating data generated on the servers
running 9.3.14 would silently fail when run on the dev VMs running 9.3.12.